Refactor(#61): 사용하지 않는 파일 제거 및 폴더 이동 - useInfiniteScroll 훅 제거 - CommentHeader 컴포넌트 제거 - BaseCommentItem 스토리북 및 스켈레톤 컴포넌트 폴더 이동 (CommentItem 폴더 => BaseCommentItem)
