+++ src/api/comments/replies.queries.ts
@@ -1,6 +1,6 @@
 import { useInfiniteQuery } from '@tanstack/react-query'
 
-import { ReplyType } from '@/types/comments/comments.type'
+import { ReplyType } from '@/types/comments'
 import { ApiResponse } from '@/types/common'
 
 import { replyKeys } from './key/replies.keys'
