+++ src/types/comments/index.ts
@@ -1,0 +1,2 @@
+export * from './comments.type'
+export * from './reply.type'
